add api_error.py
