+++ backend/ops_api/ops/auth/api.py
@@ -2,9 +2,9 @@ from flask import Response, request
 from flask_jwt_extended import jwt_required
 
 from ops_api.ops.auth import bp
+from ops_api.ops.auth.api_error import handle_api_error
 from ops_api.ops.auth.schema import LoginRequestSchema, LoginResponseSchema
 from ops_api.ops.auth.service import login, logout, refresh
-from ops_api.ops.auth.utils import handle_api_error
 from ops_api.ops.utils.errors import error_simulator
 from ops_api.ops.utils.response import make_response_with_headers
